Rename and structure tests.
