+++ test-integration/scenarios/escrow/index.js
@@ -1,9 +1,21 @@
-const _1_uniswap_swap_token_for_eth = require('./1_uniswap_swap_token_for_eth')
-const _2_uniswap_swap_token_for_token = require('./2_uniswap_swap_token_for_token')
-const _9_repay_loan_in_full = require('./9_repay_loan_in_full');
+const _1_escrow_repay_loan_in_full = require('./1_escrow_repay_loan_in_full');
+const _1_uniswap_swap_token_for_eth = require('./dapps/uniswap/1_uniswap_swap_token_for_eth')
+const _2_uniswap_swap_token_for_token = require('./dapps/uniswap/2_uniswap_swap_token_for_token')
+
+const escrow = {
+  _1_escrow_repay_loan_in_full,
+}
+
+const uniswap = {
+  _1_uniswap_swap_token_for_eth,
+  _2_uniswap_swap_token_for_token
+}
+
+const dapps = {
+  ...uniswap
+}
 
 module.exports = {
-  _1_uniswap_swap_token_for_eth,
-  _2_uniswap_swap_token_for_token,
-  _9_repay_loan_in_full,
+  ...escrow,
+  ...dapps
 }
